hw&scripts/api_final_yatube/requirements.txt
requests
django
pytest
pytest-django
djangorestframework
djangorestframework-simplejwt


pillow>=7.1.0             # images

